add argument and maintenence runs

+++ train_nets/fit_CNN.py
@@ -464,7 +464,7 @@ def load_optimizer(config, model):
             state_dict = pickle.load(f)
         optimizer.load_state_dict(state_dict)
         for g in optim.param_groups:
-            g['lr'] = 0.001
+            g['lr'] =torch.float64(0.001)
         print(type(optimizer.param_groups[0]),optimizer.param_groups[0])
         # optimizer.param_groups[0]=config.optimizer_params['lr']
     exit(0)
